job-manager: refactor events journal to new files
Refactor all job-manager events journal streaming into
new files journal.[ch].

For the most part, the code in journal.[ch] is a copy of the events
journaling / streaming from events.[ch], with adjustments to
variable / function names.
