+++ src/modules/job-manager/Makefile.am
@@ -41,7 +41,9 @@ job_manager_la_SOURCES = \
 	priority.h \
 	priority.c \
 	annotate.h \
-	annotate.c
+	annotate.c \
+	journal.h \
+	journal.c
 
 job_manager_la_LDFLAGS = $(fluxmod_ldflags) -module
 job_manager_la_LIBADD = $(fluxmod_libadd) \
@@ -69,6 +71,7 @@ test_ldadd = \
         $(top_builddir)/src/modules/job-manager/submit.o \
         $(top_builddir)/src/modules/job-manager/wait.o \
         $(top_builddir)/src/modules/job-manager/annotate.o \
+        $(top_builddir)/src/modules/job-manager/journal.o \
 	$(top_builddir)/src/common/libtap/libtap.la \
 	$(top_builddir)/src/common/libjob/libjob.la \
 	$(top_builddir)/src/common/libflux-internal.la \
