+++ src/modules/job-manager/event.c
@@ -43,28 +43,21 @@
 #include "alloc.h"
 #include "start.h"
 #include "drain.h"
+#include "journal.h"
 #include "wait.h"
 
 #include "event.h"
 
 #include "src/common/libeventlog/eventlog.h"
 
-#define EVENTS_JOURNAL_MAXLEN 1000
-
 const double batch_timeout = 0.01;
 
 struct event {
     struct job_manager *ctx;
-    flux_msg_handler_t **handlers;
     struct event_batch *batch;
     flux_watcher_t *timer;
     zlist_t *pending;
     zlist_t *pub_futures;
-    zlist_t *listeners;
-
-    /* events_journal holds most recent events for listeners */
-    zlist_t *events_journal;
-    int events_journal_maxlen;
 };
 
 struct event_batch {
@@ -72,14 +65,7 @@ struct event_batch {
     flux_kvs_txn_t *txn;
     flux_future_t *f;
     json_t *state_trans;
-    bool listener_response_available;
     zlist_t *responses; // responses deferred until batch complete
-};
-
-struct events_listener {
-    const flux_msg_t *request;
-    json_t *allow;
-    json_t *deny;
 };
 
 static struct event_batch *event_batch_create (struct event *event);
@@ -132,9 +118,7 @@ static void event_batch_commit (struct event *event)
         /* note that job-state events will be sent after the KVS
          * commit, as we want to ensure anyone who receives a
          * job-state transition event will be able to read the
-         * corresponding event in the KVS.  The event stream does not
-         * have such a requirement, since we're sending the listener
-         * the data.
+         * corresponding event in the KVS.
          */
         if (batch->txn) {
             if (!(batch->f = flux_kvs_commit (ctx->h, NULL, 0, batch->txn)))
@@ -527,104 +511,6 @@ static int get_timestamp_now (double *timestamp)
     return 0;
 }
 
-/* wrap the eventlog entry in another object with the job id and
- * eventlog_seq.
- *
- * The job id is necessary so listeners can determine which job the
- * event is associated with.
- *
- * The eventlog sequence number is necessary so users can determine if the
- * event is a duplicate if they are reading events from another source
- * (i.e. they could be reading events from the job's eventlog in the
- * KVS).
- */
-static json_t *wrap_events_entry (flux_jobid_t id,
-                                  int eventlog_seq,
-                                  json_t *entry)
-{
-    json_t *wrapped_entry;
-    if (!(wrapped_entry = json_pack ("{s:I s:i s:O}",
-                                     "id", id,
-                                     "eventlog_seq", eventlog_seq,
-                                     "entry", entry))) {
-        errno = ENOMEM;
-        return NULL;
-    }
-    return wrapped_entry;
-}
-
-static bool allow_deny_check (struct events_listener *el, const char *name)
-{
-    bool add_entry = true;
-
-    if (el->allow) {
-        add_entry = false;
-        if (json_object_get (el->allow, name))
-            add_entry = true;
-    }
-
-    if (add_entry && el->deny) {
-        if (json_object_get (el->deny, name))
-            add_entry = false;
-    }
-
-    return add_entry;
-}
-
-static void json_decref_wrapper (void *data)
-{
-    json_t *o = (json_t *)data;
-    json_decref (o);
-}
-
-int event_batch_process_event_entry (struct event *event,
-                                     flux_jobid_t id,
-                                     int eventlog_seq,
-                                     const char *name,
-                                     json_t *entry)
-{
-    struct events_listener *el;
-    json_t *wrapped_entry = NULL;
-    int saved_errno;
-
-    if (!(wrapped_entry = wrap_events_entry (id,
-                                             eventlog_seq,
-                                             entry)))
-        goto error;
-
-    el = zlist_first (event->listeners);
-    while (el) {
-        if (allow_deny_check (el, name)) {
-            if (flux_respond_pack (event->ctx->h, el->request,
-                                   "{s:[O]}", "events", wrapped_entry) < 0)
-                flux_log_error (event->ctx->h, "%s: flux_respond_pack",
-                                __FUNCTION__);
-        }
-
-        el = zlist_next (event->listeners);
-    }
-
-    if (zlist_size (event->events_journal) > event->events_journal_maxlen)
-        zlist_remove (event->events_journal, zlist_head (event->events_journal));
-    if (zlist_append (event->events_journal, json_incref (wrapped_entry)) < 0)
-        goto nomem;
-    zlist_freefn (event->events_journal,
-                  wrapped_entry,
-                  json_decref_wrapper,
-                  true);
-
-    json_decref (wrapped_entry);
-    return 0;
-
-nomem:
-    errno = ENOMEM;
-error:
-    saved_errno = errno;
-    json_decref (wrapped_entry);
-    errno = saved_errno;
-    return -1;
-}
-
 int event_job_post_pack (struct event *event,
                          struct job *job,
                          const char *name,
@@ -645,11 +531,11 @@ int event_job_post_pack (struct event *event,
     if (!(entry = eventlog_entry_vpack (timestamp, name, context_fmt, ap)))
         return -1;
     /* call before eventlog_seq increment below */
-    if (event_batch_process_event_entry (event,
-                                         job->id,
-                                         eventlog_seq,
-                                         name,
-                                         entry) < 0)
+    if (journal_process_event (event->ctx->journal,
+                               job->id,
+                               eventlog_seq,
+                               name,
+                               entry) < 0)
         goto error;
     if ((flags & EVENT_JOURNAL_ONLY))
         goto out;
@@ -686,234 +572,12 @@ error:
     return -1;
 }
 
-static void events_listener_destroy (void *data)
-{
-    struct events_listener *el = (struct events_listener *)data;
-    if (el) {
-        int saved_errno = errno;
-        flux_msg_decref (el->request);
-        json_decref (el->allow);
-        json_decref (el->deny);
-        free (el);
-        errno = saved_errno;
-    }
-}
-
-static struct events_listener *events_listener_create (const flux_msg_t *msg,
-                                                       json_t *allow,
-                                                       json_t *deny)
-{
-    struct events_listener *el;
-
-    if (!(el = calloc (1, sizeof (*el))))
-        goto error;
-    el->request = flux_msg_incref (msg);
-    el->allow = json_incref (allow);
-    el->deny = json_incref (deny);
-    return el;
-error:
-    events_listener_destroy (el);
-    return NULL;
-}
-
-static void events_handle_request (flux_t *h,
-                                   flux_msg_handler_t *mh,
-                                   const flux_msg_t *msg,
-                                   void *arg)
-{
-    struct job_manager *ctx = arg;
-    struct event *event = ctx->event;
-    struct events_listener *el = NULL;
-    const char *errstr = NULL;
-    json_t *allow = NULL;
-    json_t *deny = NULL;
-    json_t *a = NULL;
-    json_t *wrapped_entry;
-
-    if (flux_request_unpack (msg, NULL, "{s?o s?o}",
-                             "allow", &allow,
-                             "deny", &deny) < 0)
-        goto error;
-
-    if (!flux_msg_is_streaming (msg)) {
-        errno = EPROTO;
-        errstr = "job-manager.events requires streaming RPC flag";
-        goto error;
-    }
-
-    if (allow && !json_is_object (allow)) {
-        errno = EPROTO;
-        errstr = "job-manager.events allow should be an object";
-        goto error;
-    }
-
-    if (deny && !json_is_object (deny)) {
-        errno = EPROTO;
-        errstr = "job-manager.events deny should be an object";
-        goto error;
-    }
-
-    if (!(el = events_listener_create (msg, allow, deny)))
-        goto error;
-
-    if (zlist_append (event->listeners, el) < 0) {
-        errno = ENOMEM;
-        goto error;
-    }
-    zlist_freefn (event->listeners, el, events_listener_destroy, true);
-
-    if (event_batch_start (event) < 0)
-        goto error;
-
-    wrapped_entry = zlist_first (event->events_journal);
-    while (wrapped_entry) {
-        const char *name;
-
-        if (json_unpack (wrapped_entry,
-                         "{s:{s:s}}",
-                         "entry",
-                           "name", &name) < 0) {
-            flux_log (ctx->h, LOG_ERR, "invalid wrapped entry");
-            goto error;
-        }
-
-        if (allow_deny_check (el, name)) {
-            if (!a) {
-                if (!(a = json_array ()))
-                    goto nomem;
-            }
-            if (json_array_append (a, wrapped_entry) < 0)
-                goto nomem;
-        }
-        wrapped_entry = zlist_next (event->events_journal);
-    }
-
-    if (a && json_array_size (a) > 0) {
-        if (flux_respond_pack (ctx->h, el->request,
-                               "{s:O}", "events", a) < 0) {
-            flux_log_error (ctx->h, "%s: flux_respond_pack",
-                            __FUNCTION__);
-            goto error;
-        }
-    }
-
-    json_decref (a);
-    return;
-
-nomem:
-    errno = ENOMEM;
-error:
-    if (flux_respond_error (h, msg, errno, errstr) < 0)
-        flux_log_error (h, "%s: flux_respond_error", __FUNCTION__);
-    events_listener_destroy (el);
-    json_decref (a);
-}
-
-static bool match_events_listener (struct events_listener *el,
-                                   uint32_t matchtag,
-                                   const char *sender)
-{
-    uint32_t t;
-    char *s = NULL;
-    bool found = false;
-
-    if (!flux_msg_get_matchtag (el->request, &t)
-        && matchtag == t
-        && !flux_msg_get_route_first (el->request, &s)
-        && !strcmp (sender, s))
-        found = true;
-    free (s);
-    return found;
-}
-
-static void events_cancel_request (flux_t *h, flux_msg_handler_t *mh,
-                                   const flux_msg_t *msg, void *arg)
-{
-    struct job_manager *ctx = arg;
-    struct event *event = ctx->event;
-    struct events_listener *el;
-    uint32_t matchtag;
-    char *sender = NULL;
-
-    if (flux_request_unpack (msg, NULL, "{s:i}", "matchtag", &matchtag) < 0
-        || flux_msg_get_route_first (msg, &sender) < 0) {
-        flux_log_error (h, "error decoding events-cancel request");
-        return;
-    }
-    el = zlist_first (event->listeners);
-    while (el) {
-        if (match_events_listener (el, matchtag, sender))
-            break;
-        el = zlist_next (event->listeners);
-    }
-    if (el) {
-        if (flux_respond_error (h, el->request, ENODATA, NULL) < 0)
-            flux_log_error (h, "%s: flux_respond_error", __FUNCTION__);
-        zlist_remove (event->listeners, el);
-    }
-    free (sender);
-}
-
-static int create_zlist_and_append (zlist_t **lp, void *item)
-{
-    if (!*lp && !(*lp = zlist_new ())) {
-        errno = ENOMEM;
-        return -1;
-    }
-    if (zlist_append (*lp, item) < 0) {
-        errno = ENOMEM;
-        return -1;
-    }
-    return 0;
-}
-
-void event_listeners_disconnect_rpc (flux_t *h,
-                                     flux_msg_handler_t *mh,
-                                     const flux_msg_t *msg,
-                                     void *arg)
-{
-    struct job_manager *ctx = arg;
-    struct event *event = ctx->event;
-    struct events_listener *el;
-    char *sender;
-    zlist_t *tmplist = NULL;
-
-    if (flux_msg_get_route_first (msg, &sender) < 0)
-        return;
-    el = zlist_first (event->listeners);
-    while (el) {
-        char *tmpsender;
-        if (flux_msg_get_route_first (el->request, &tmpsender) == 0) {
-            if (!strcmp (sender, tmpsender)) {
-                /* cannot remove from zlist while iterating, so we
-                 * store off entries to remove on another list */
-                if (create_zlist_and_append (&tmplist, el) < 0) {
-                    flux_log_error (h, "job-manager.disconnect: "
-                                    "failed to remove event listener");
-                    free (tmpsender);
-                    goto error;
-                }
-            }
-            free (tmpsender);
-        }
-        el = zlist_next (event->listeners);
-    }
-    if (tmplist) {
-        while ((el = zlist_pop (tmplist)))
-            zlist_remove (event->listeners, el);
-    }
-    free (sender);
-error:
-    zlist_destroy (&tmplist);
-}
-
 /* Finalizes in-flight batch KVS commits and event pubs (synchronously).
  */
 void event_ctx_destroy (struct event *event)
 {
     if (event) {
         int saved_errno = errno;
-        flux_msg_handler_delvec (event->handlers);
         flux_watcher_destroy (event->timer);
         event_batch_commit (event);
         if (event->pending) {
@@ -932,51 +596,18 @@ void event_ctx_destroy (struct event *event)
             }
         }
         zlist_destroy (&event->pub_futures);
-        if (event->listeners) {
-            struct events_listener *el;
-            while ((el = zlist_pop (event->listeners))) {
-                if (flux_respond_error (event->ctx->h,
-                                        el->request,
-                                        ENODATA, NULL) < 0)
-                    flux_log_error (event->ctx->h, "%s: flux_respond_error",
-                                    __FUNCTION__);
-                events_listener_destroy (el);
-            }
-            zlist_destroy (&event->listeners);
-        }
-        if (event->events_journal)
-            zlist_destroy (&event->events_journal);
         free (event);
         errno = saved_errno;
     }
 }
 
-static const struct flux_msg_handler_spec htab[] = {
-    {
-        FLUX_MSGTYPE_REQUEST,
-        "job-manager.events",
-        events_handle_request,
-        0
-    },
-    {
-        FLUX_MSGTYPE_REQUEST,
-        "job-manager.events-cancel",
-        events_cancel_request,
-        0
-    },
-    FLUX_MSGHANDLER_TABLE_END,
-};
-
 struct event *event_ctx_create (struct job_manager *ctx)
 {
     struct event *event;
-    flux_conf_error_t err;
 
     if (!(event = calloc (1, sizeof (*event))))
         return NULL;
     event->ctx = ctx;
-    if (flux_msg_handler_addvec (ctx->h, htab, ctx, &event->handlers) < 0)
-        goto error;
     if (!(event->timer = flux_timer_watcher_create (flux_get_reactor (ctx->h),
                                                     0.,
                                                     0.,
@@ -987,22 +618,6 @@ struct event *event_ctx_create (struct job_manager *ctx)
         goto nomem;
     if (!(event->pub_futures = zlist_new ()))
         goto nomem;
-    if (!(event->listeners = zlist_new ()))
-        goto nomem;
-    if (!(event->events_journal = zlist_new ()))
-        goto nomem;
-    event->events_journal_maxlen = EVENTS_JOURNAL_MAXLEN;
-
-    if (flux_conf_unpack (flux_get_conf (ctx->h),
-                          &err,
-                          "{s?{s?i}}",
-                          "job-manager",
-                            "events_journal_maxlen",
-                            &event->events_journal_maxlen) < 0) {
-        flux_log (ctx->h, LOG_ERR,
-                  "error reading job-manager config: %s",
-                  err.errbuf);
-    }
 
     return event;
 nomem:
@@ -1010,13 +625,6 @@ nomem:
 error:
     event_ctx_destroy (event);
     return NULL;
-}
-
-int event_listeners_count (struct event *event)
-{
-    if (event)
-        return zlist_size (event->listeners);
-    return -1;
 }
 
 /*
